Update to subroutines/omni_compute_ebars.pro
Update error-bar computation routine to gracefully handle bimodal PDFs computed
from Monte Carlo calculations of derived quantities.  Sometimes the secondary
peak (with ~15-20% of the probability) can have a taller peak in the PDF because
of binning considerations, but this peak is not the one we want to select as
the "Maximum-Likelihood" peak because the peak itself does not contain a
majority of the probability.  Updates to the routine check for bimodality, and
properly mask the PDF, if desired.