Trigonometry Fix in rot_curve/
Fix a long-standing issue in rot_curve/omni_lbd2rz.pro and
rot_curve/omni_vphys2vlsr.pro that required ellaborate workarounds to make the
routines spit out the correct values.  The issue was in the computation of
Galactocentric azimuth in OMNI_LBD2RZ, which used an ACOS() call -- but arc
cosine only returns 180 degrees worth of angle.  Switched to an ATAN() call
which correctly identifies the Galactocentric quadrant.

With this correction, the kludge in computing PHI (angle between velocity and
line of sight) in OMNI_VPHYS2VLSR() was removed, leaving the (long-ago) original
version of the equation.