Add parallax parser routine
Routine parses the Table 1 of Reid et al. (2014) of parallax data for 103
sources across the Galactic plane.  That table is listed in celestial (RA,Dec)
coordinates (honestly, when dealing with Galactic plane stuff, WHY??!?), and
this routine computes (l,b), as well as saves the table into an IDL save file.