Add LOCAL conffile element OUTPUT
Add OUTPUT keyword to the local_layout.conf configuration file.  This allows the
user to create an output directory in the location of their choosing and tell
the configuration file where that is.  This obviates the need to have an output
directory in the local/ directory -- or have a symlink at that location.

This change requires updates to several files:
	modified:   conffiles/local_layout.conf
	modified:   distance_omnibus.pro
	modified:   emaf/omni_emaf_morph.pro
	modified:   emaf/omni_mir_examine.pro
	modified:   subroutines/omni_read_conffile_local.pro

Additionally, the routine emaf/driver_morph.pro is removed from the repository
because it (1) has not been updated to the new omni_* architecture, and (2) its
usefulness has been replaced by the LOCAL.MAKEEMAFPS flag.  The emaf/README has
been updated to reflect this deletion.