Check for and define !CONST system variable
IDL v8.2 introduced the !CONST system variable which contains all sorts of very
useful physical constants, all in double precision.  Since we assume an IDL
version 8.0+ for the running of distance-omnibus, a compatibility routine is
required to check for the existance of !CONST, and define it if not present.

This compatibility routine is utils/compat_const.pro, and the other routines in
the repository that use !CONST have been modified to call this new routine right
off the bat.

Note: The version of !CONST included in utils/compat_const.pro is the one from
IDL v8.2.3; if significant changes to the IDL-supplied version of !CONST occur,
this this routine will need to be updated if future revisions of
distance-omnibus make use of the new features.