Add 'All Done!' message
Add an "all done" message to the end of distance_omnibus.pro so that users know
that it ran successfully and that it didn't crash.  Also, removed the 'help'
output at the end of subroutines/omni_export_fits.pro to avoid confusion (i.e.
the code outputs technical stuff when it pukes).