Name conflict issue resolved (gauss2 --> guass_1)
In the Markwardt library, there is a routine gauss1.pro, which creates a 1-D
Gaussian based on an x-array and [amplitude,mean,sigma], but the order of that
array is screwy.  So, I had created my own "gauss2.pro" (increment, right?) to
use the usual output of the A array from MPFITPEAK as the input.

So, it turns out there is a gauss2.pro routine in the Markwardt library to
create a 2-D Gaussian.  If the Markwardt library is _earlier_ in the IDL search
path, then that routine is chosen, causing issues.

So, a simple rename of my hacked gauss2.pro --> gauss_1.pro should make everyone
happy, no matter in what order the IDL path is specified.  The affected
routines are listed below for reference.

	modified:   dpdf/prob_hrds.pro
	modified:   dpdf/prob_knownd.pro
	modified:   dpdf/prob_parallax.pro
	modified:   subroutines/fit_sa_spectrum.pro
	modified:   subroutines/omni_vspec.pro
	modified:   subroutines/sa_gauss.pro
	modified:   subroutines/velocity_fit_spectrum.pro
	deleted:    utils/gauss1.pro
	renamed:    utils/gauss2.pro -> utils/gauss_1.pro