Check for IDL version >= 8.0
Add check for "recent" version of IDL, since some aspects of the code rely on
constructs introduced in or after this version.  Version 8.0 has been around
long enough (July 2010) that its requirement should not be a burden on those
wanting to use the distance-omnibus code base.

Check routine added, and modifications done to "front-line" routines to call
the check routine.  If an earler version of IDL is detected, execution halts.

	new file:   subroutines/omni_check_version.pro

	modified:   distance_omnibus.pro
	modified:   create_local_files/omni_assoc_catalog.pro
	modified:   create_local_files/omni_glimpse_emaf.pro
	modified:   create_local_files/omni_glimpse_postage.pro
	modified:   create_local_files/omni_glimpse_starsub.pro
	modified:   create_local_files/omni_grs_spectra_local.pro
	modified:   create_local_files/omni_import_velocity.pro
	modified:   create_local_files/omni_survey_postage.pro