Convert storage of map information into lists (requires IDL 8.0+)
Convert storage of map information from individual named variables into lists
(requires IDL 8.0+) to avoid excessive use of the Execute() command and the
feeling of spaghetti code.

Also, for the SURVEY structure, tabulate the number of pixels in each source,
and the peak flux density pixel value.  These are to be used with the surface
mass density computations.

	modified:   create_local_files/omni_assoc_catalog.pro