Enhancement to plotting routine, cosmetic fix for KDIST DPDFs
Enhancement for the plotting routine omni_plot_dpdfs.pro to allow for user
choice in the look of the output.

Cosmetic fix for the subroutine in prob_kdist.pro which squashes the extraneous
probabilites for objects beyond the Solar Circle -- the otherwise-uniform
probability for sources with no dense gas vLSR was being modified by the
squasher.  This is of no consequence except for plotting of DPDFs with
omni_plot_dpdfs.pro, which can lead to some confusing plots.

	modified:   dpdf/prob_kdist.pro
	modified:   subroutines/omni_plot_dpdfs.pro