Correct fix for the "Big Island" problem in OMNI_GRS_MASKSPEC
Correct fix for the "Big Island" problem, which was attempted in the 09/17/14
fix.  Now, the routine identifies islands above the THRESHOLD value, then
analyzes each island for multiple summits (a la the Big Island of Hawaii).  In
this way, if one island has multiple summits, they will be analyzed
individually using the criteria for multiple peaks in the spectrum.

	modified:   subroutines/omni_grs_maskspec.pro