Add dV/dR rotation curve slope option
Add the dV/dR rotation curve parameter from Reid et al. (2014) to the kinematic
distance calculation and configuration files.  For the moment, we are using
dV/dR = 0, which is consistent with the measured -0.2 +- 0.4 km/s/kpc.  Adding
this parameter to the calculation, however, allows for the easy use of a nonzero
value of this parameter with future revisions of this rotation curve.

	modified:   conffiles/galactic_params.conf
	modified:   rot_curve/omni_vphys2vlsr.pro
	modified:   subroutines/omni_read_conffile_galactic.pro