Correct convention that V_s is in the direction of circular motion
Update kinematic distance routines with the convention that V_s, as specified
in Reid et al. (2014) is in the direction of circular motion.  Therefore, any
obserbved counterrotation would have negative values of V_s.  The code
previously assumed a counterrotation, and hard-wired the negative sign.  With
this update, routines have been aligned with the current convention.

	modified:   rot_curve/kdist.pro
	modified:   rot_curve/kdist_iau.pro
	modified:   rot_curve/omni_vphys2vlsr.pro